// frontend/src/api/model/getFoodsResponseDto.ts
/**
 * Generated by orval v8.14.0 🍺
 * Do not edit manually.
 * YaoYao Dinner API
 * Restaurant management API
 * OpenAPI spec version: 1.0
 */
import type { FoodItemDto } from './foodItemDto';

export interface GetFoodsResponseDto {
  foods: FoodItemDto[];
  page: number;
  count: number;
  total: number;
}